reorganize code and update comments
